Add a new header field to keep track of the last point in time a package was reviewed : %LASTCHECKED
While several packages we use aren't updated often, the few that are, such as auctex, magit, etc. are hard to stay up to date with without running the risk of missing some, potentially crucial informations.
This header property tracks the last commit of a package we checked.
This assumes we're familiar of the history of the package before the commit, or, at least, of the state of the package at the time of the commit.
Indeed, the goal is to alleviate the reviewing time, by making it so we only have to look at what's new rather than reread the whole repo; however, we still need to do this first full reading the first time we set the property from None to a commit.

There are tools such as :pin and lock-files to lock our init to a specific state, however, lock-files seems to be designed backwards to me.
Indeed, it saves the current state to a file and lock all of the versions, instead of asking for what to lock, what to look for, etc.
Furthermore, it means that ensuring two system are on the same versions, just from a lock-file isn't evident, since it is another file to track, and the first install of all the packages doesn't seem to adhere to the values from what the issues of elpaca seem to indicate. NOTE that it may have since changed !
As for :pin, in the end, it is fine if we pull from commits we haven't reviewed yet, especially since it may be a cause of greater pain when installing on a new system or refreshing the install.