Dashboard : create a folder for banners + read commit message for suggestion.
Firstly, the old code used to call some hook twice, due to poor dashboard documentation,
see  https://github.com/emacs-dashboard/emacs-dashboard/issues/499
Secondly, the code didn't properly specify random banner selection, it's now fixed, requires a setq in init;
we could technically remove the setq, put it back in custom, as long as we assume all the files in the banner are of a correct format.
This would actually be a good github CL thing.

We still need to do the buttons, planning on : github repos, gitgay repos, pink-bliss-uwu, init.org

We need to decide on whether we include icons or not, all-the-icons and nerd-icons are supported, but they seem like very heavy packages, especially if we only use them with dashboard.